development/requirements.txt
# local package
-e .

pandas==1.1.1
numpy==1.18.1
scikit-learn==0.22.1
scipy
catboost==0.24.1
jupyter
xlrd # for pd.read_excel
joblib==0.14.1

# external requirements
click
python-dotenv>=0.5.1

